Move prompt from customizations and add arguments to cli.json
