+++ awscli/data/cli.json
@@ -77,6 +77,14 @@
         "no-cli-pager": {
             "action": "store_true",
             "help": "<p>Disable cli pager for output.</p>"
+        },
+        "cli-auto-prompt": {
+            "action": "store_true",
+            "help": "<p>Automatically prompt for CLI input parameters.</p>"
+        },
+        "no-cli-auto-prompt": {
+            "action": "store_true",
+            "help": "<p>Disable automatically prompt for CLI input parameters.</p>"
         }
     }
 }
